got notes buttons to render, list not rendering yet
