+++ client/src/App.js
@@ -1,12 +1,12 @@
 import React, { Component } from 'react';
 import './App.css';
-import Header from './components/panel';
+import SideBar from './components/sideBar';
 
 class App extends Component {
   render() {
     return (
       <div className="App">
-        <Header />
+        <SideBar />
       </div>
     );
   }
