+++ client/src/index.css
@@ -3,3 +3,5 @@ body {
   padding: 0;
   font-family: sans-serif;
 }
+
+@import url('https://fonts.googleapis.com/css?family=Raleway|Roboto:700');
